+++ modules/framegrabber/include/ifm3d/fg/frame.h
@@ -174,6 +174,10 @@ namespace ifm3d
     */
     O3R_RESULT_IMU = static_cast<uint64_t>(ifm3d::image_chunk::O3R_RESULT_IMU),
 
+    /** @hideinitializer
+     * @brief ifm3d::ODSPolarOccupancyGridV1
+    */
+    O3R_ODS_POLAR_OCC_GRID = static_cast<uint64_t>(ifm3d::image_chunk::O3R_ODS_POLAR_OCC_GRID),
 
     /** @hideinitializer 
      * @brief The point cloud encoded as a 3 channel XYZ image