+++ ChangeLog.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added missing Python binding methods
 - Added support for Python 3.13 and dropped support for Python `3.8.10`
 - Added support for Ubuntu `24.04` and dropped support for Ubuntu `18.04`
+- Add Deserialize module
+  - Add struct `ODSPolarOccupancyGridV1`
 
 ## 1.6.3 - 2024-11-29